refactor: migrate museum-related types and queries to shared directory
