+++ src/pages/admin/AdminMuseums.tsx
@@ -2,14 +2,14 @@ import { useState } from "react";
 import { ColumnDef } from "@tanstack/react-table";
 import { Link } from "react-router-dom";
 import { DataTable } from "@/shared/components/DataTable";
-import { Museum } from "@/types/Museums";
+import { Museum } from "@/shared/types/museums.types";
 import PhotoCellModal from "@/shared/components/PhotoCellModal";
 import { createMuseum, deleteMuseum, editMuseum } from "@/services/Museums";
 import { useToast } from "@/hooks/use-toast";
 import MuseumForm from "@/features/admin/components/MuseumForm";
 import HoursDisplay from "@/features/admin/components/HoursDisplay";
 import { Clock, MapPin } from "lucide-react";
-import { MuseumHours } from "@/types/Museums";
+import { MuseumHours } from "@/shared/types/museums.types";
 import { useFetchMuseums } from "@/features/admin/queries/useMuseumsQuery";
 import Loader from "@/shared/components/Loader";
 import { AxiosError } from "axios";
