+++ src/pages/admin/AdminTours.tsx
@@ -7,10 +7,10 @@ import { deleteTour, editTour } from "@/services/tour.services";
 import { getAllMuseums } from "@/services/Museums";
 import { useToast } from "@/hooks/use-toast";
 import { Link } from "react-router-dom";
-import { Museum } from "@/types/Museums";
+import { Museum } from "@/shared/types/museums.types";
 import TourForm from "@/features/admin/components/TourForm";
 import { TagsCell } from "@/shared/components/TagCell";
-import { useCreateTour, useFetchTours } from "@/querys/tour.querys";
+import { useCreateTour, useFetchTours } from "@/features/tour/tour.querys";
 import Loader from "@/shared/components/Loader";
 import { Tag } from "@/types/tag";
 import { Badge } from "@/components/ui/badge";
